Change CasperPacketHandler to be just... PacketHandler
AND NOTHING ELSE!
